requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bitarray==3.8.0
certifi==2026.1.4
charset-normalizer==3.4.4
ckzg==2.1.5
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==6.0.0b1
frozenlist==1.8.0
gql==4.0.0
graphql-core==3.2.7
h11==0.16.0
h2==4.3.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
lomond==0.3.3
multidict==6.7.0
packaging==25.0
parsimonious==0.10.0
pluggy==1.6.0
poly_eip712_structs==0.0.1
polymarket-apis==0.4.3
propcache==0.4.1
py_builder_signing_sdk==0.0.2
py_clob_client==0.34.5
py_order_utils==0.3.2
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyunormalize==17.0.0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
six==1.17.0
toolz==1.1.0
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
web3==7.14.0
websocket-client==1.9.0
websockets==15.0.1
wsaccel==0.6.7
yarl==1.22.0
